+++ src/generate_dataset_docs.py
@@ -4,10 +4,13 @@ import argparse
 from pathlib import Path
 import sys
 
+from tableschema_to_template.create_xlsx import create_xlsx
+
 from ingest_validation_tools.schema_loader import (
     list_types, get_table_schema, get_directory_schema)
 from ingest_validation_tools.docs_utils import (
-    get_tsv_name, generate_template_tsv, generate_readme_md)
+    get_tsv_name, get_xlsx_name,
+    generate_template_tsv, generate_readme_md)
 from ingest_validation_tools.argparse_types import dir_path
 
 
@@ -26,10 +29,14 @@ def main():
     table_schema = get_table_schema(args.type)
     directory_schema = get_directory_schema(args.type)
 
-    with open(Path(args.target) / get_tsv_name(args.type), 'w') as f:
-        f.write(generate_template_tsv(table_schema))
+    # README:
     with open(Path(args.target) / 'README.md', 'w') as f:
         f.write(generate_readme_md(table_schema, directory_schema, args.type))
+
+    # Data entry templates:
+    with open(Path(args.target) / get_tsv_name(args.type), 'w') as f:
+        f.write(generate_template_tsv(table_schema))
+    create_xlsx(table_schema, Path(args.target) / get_xlsx_name(args.type))
 
 
 if __name__ == "__main__":