+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Warn if two TSVs are for the same assay type.
 - Add SLIDEseq
 - Add antibodies.tsv.
+- Generate Excel files.
 
 ## v0.0.5 - 2020-11-09
 - Change "mixif" to "mxif".